refactor: move create-ticket to top-level
