+++ cmd/kubectl-testkube/commands/createticket.go
@@ -1,8 +1,9 @@
-package debuginfo
+package commands
 
 import (
 	"github.com/kubeshop/testkube/cmd/kubectl-testkube/commands/common"
-	"github.com/kubeshop/testkube/cmd/kubectl-testkube/commands/debug/github"
+	debuginfo "github.com/kubeshop/testkube/cmd/kubectl-testkube/commands/debug"
+	"github.com/kubeshop/testkube/cmd/kubectl-testkube/commands/github"
 	"github.com/kubeshop/testkube/pkg/ui"
 	"github.com/spf13/cobra"
 )
@@ -14,7 +15,7 @@ func NewCreateTicketCmd() *cobra.Command {
 		Long:  "Create an issue of type bug in the Testkube repository",
 		Run: func(cmd *cobra.Command, args []string) {
 			client, _ := common.GetClient(cmd)
-			debug, err := getDebugInfo(client)
+			debug, err := debuginfo.GetDebugInfo(client)
 			ui.ExitOnError("get debug info", err)
 			ui.ExitOnError("opening GitHub ticket", github.OpenTicket(debug))
 		},
