+++ docs/cli/kubectl-testkube_create_ticket.md
